Design and Order of online estimate correction
Now the order of the pages will appear in the correct order. onlineEstimates will arrive the user at the page to input there information. OnlineEstimates/vehicle will arive at vehicle infomation page
